beetle-quest/tests/locust/requirements.txt
blinker==1.9.0
Brotli==1.1.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
ConfigArgParse==1.7
cryptography==43.0.3
Flask==3.1.0
Flask-Cors==5.0.0
Flask-Login==0.6.3
gevent==24.11.1
geventhttpclient==2.3.1
greenlet==3.1.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
locust==2.32.3
MarkupSafe==3.0.2
msgpack==1.1.0
psutil==6.1.0
pycparser==2.22
PyJWT==2.10.0
pyotp==2.9.0
pyzmq==26.2.0
requests==2.32.3
setuptools==75.5.0
urllib3==2.2.3
Werkzeug==3.1.3
zope.event==5.0
zope.interface==7.1.1
